pi/requirements.txt
numpy
scipy
pyserial
fastapi
uvicorn[standard]
websockets
pyyaml
pytest
tflite-runtime
resampy
